+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Fix `__init__` in subclasses of protocols.
 - Fix incorrect behaviour on Python 3.9 and Python 3.10 that meant that
   calling `isinstance` with `typing_extensions.Concatenate[...]` or
   `typing_extensions.Unpack[...]` as the first argument could have a different